Clean up code
still doesn't work with inequalities yet.
